fix: calendar exporting non-page functions
